+++ CHANGELOG.md
@@ -6,7 +6,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/)
 with respect to its command line interface and HTTP interface.
 
-## [Unreleased](//github.com/opentable/sous/compare/0.5.51...HEAD)
+## [0.5.53](//github.com/opentable/sous/compare/0.5.52...0.5.53)
+
+### Fixed
+* Server: the resolve complete message is properly emitted and produces stats
+
+## [0.5.52](//github.com/opentable/sous/compare/0.5.51...0.5.52)
+
+### Changed
+* All: Logging to Kafka no longer goes through logrus.
 
 ### Added
 * Client: console output related to deploys. Reports the number of instances